NETwork/requirements.txt

azure-core
azure-search-documents
azure-storage-blob
azure-identity
azure-ai-projects
openai
python-dotenv
PyGithub
requests
pandas
fastapi
